requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0